Shrink the shared GPUI test helper surface
The blanket `dead_code` expectation in `tests/common/mod.rs` masked unused
helpers across all forty-one integration-test binaries that include it. Reduce
that surface and move the residual suppression off the module root.

Narrowing the expectation to individual items is not possible here: each binary
is a separate crate, so a per-item `#[expect(dead_code)]` is unfulfilled in the
binaries that do use the item and raises `unfulfilled_lint_expectations` there
instead. `#[allow(dead_code)]` would sidestep that, but
`clippy::allow_attributes` is denied workspace-wide. Reduce the masked surface
instead:

- Delete `require_last_canvas_click` and `require_canvas_click_changed`, which
  were unused by every binary.
- Move `MIN_CANVAS_HEIGHT_PX` into `gpui_shell_canvas_layout`, and
  `simulate_selection_undo`/`simulate_selection_redo` into
  `gpui_history_selection_history`, each their only consumer.
- Move the remaining multi-consumer helpers into `common::shared_helpers`, which
  carries the module-scoped expectation and documents why it cannot be narrowed.
  `tests/common/mod.rs` is now a re-export facade with no suppression.
- Gate `mod common;` in `gpui_shell_resize_borders` on the same
  `not(target_os = "linux")` predicate as its uses, so the module is not compiled
  where nothing references it.
